fix a bug
I found that sometimes the program will read the log before the scan
ends. It will cause the program can not read the full scan results.So I
let the program to monitor the process of scanning which can ensure the
program reads the log only after the scan ends.